+++ src/features/noteDetails/NoteDetails.styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet, Dimensions } from 'react-native'
-import { colors } from '@/theme/theme'
+import { colors } from '@/theme/index'
 const { width } = Dimensions.get('window')
 
 export const styles = StyleSheet.create({
